Changes
what-we-do - the order of icons should be
1. Tech Ecosystems 4. Tech Solutions
so as to align with all other areas we have listed the 4 programmes and the flow within which a beneficiary has with the organisation
Our Impact - Startups = 500+ jobs & R150m+ funds raised
Tech Skills - There is also no mention that call are currently closed - please register with chatbot to be notified when calls open
Tech-Start-ups - the text in red has nothing to do with startups - please replace.
Crop venture builder image to just have headings and no text underneath the headings + change Stages to Support
There is also no mention that call are currently closed - please register with chatbot to be notified when calls open
Tech-Solutions - Our experienced developers work with start-ups and coding graduates to innovate digital solutions for within industry challenges for public and private sector organisations.
Partners - Activity partner image crops out a number of our stakeholders logos, need to rectify.
 News - can we implement the categories we spoke about to make finding articles easier e.g. Events, Skills, StartUps, Solutions, General
About Us or Contact Us - we could include the new image which outlines mLab locations - Bridging the Digital Divide image attached.